Fix 'conversion from 'int' to 'float', possible loss of data' MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@31682 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/vrt/vrtsources.cpp
@@ -2204,10 +2204,10 @@ CPLErr VRTComplexSource::RasterIOInternal( int nReqXOff, int nReqYOff,
                 }
 
                 if (nLUTItemCount)
-                    fResult = (float) LookupValue( fResult );
+                    fResult = static_cast<float>(LookupValue( fResult ));
 
                 if( nMaxValue != 0 && fResult > nMaxValue )
-                    fResult = nMaxValue;
+                    fResult = static_cast<float>(nMaxValue);
 
                 if( eBufType == GDT_Byte )
                     *pDstLocation = (GByte) MIN(255,MAX(0,fResult + 0.5));
@@ -2239,13 +2239,13 @@ CPLErr VRTComplexSource::RasterIOInternal( int nReqXOff, int nReqYOff,
             }
             else
             {
-                float fResult = (float) dfScaleOff;
+                float fResult = static_cast<float>(dfScaleOff);
 
                 if (nLUTItemCount)
-                    fResult = (float) LookupValue( fResult );
+                    fResult = static_cast<float>(LookupValue( fResult ));
 
                 if( nMaxValue != 0 && fResult > nMaxValue )
-                    fResult = nMaxValue;
+                    fResult = static_cast<float>(nMaxValue);
 
                 if( eBufType == GDT_Byte )
                     *pDstLocation = (GByte) MIN(255,MAX(0,fResult + 0.5));
